+++ odata4j-core/src/org/odata4j/expression/IntegralLiteral.java
@@ -2,5 +2,5 @@ package org.odata4j.expression;
 
 public interface IntegralLiteral extends LiteralExpression {
 
-    public abstract long getValue();
+    public abstract int getValue();
 }
